optimize(sampling): optimized the sampling to single feed-forward step

+++ common/load_data_hm36.py
@@ -228,6 +228,6 @@ class Fusion(data.Dataset):
         bb_box = np.array([0, 0, 1, 1])
         input_2D_update = input_2D
 
-        scale = np.float(1.0)
+        scale = np.float64(1.0)
 
         return cam, gt_3D, input_2D_update, action, subject, scale, bb_box, cam_ind
